initial commit
I extracted this out of MIRA because anything with a standalone systemd service is not a microservice. This repository is NOT ready for primetime because there are a bunch of bugfixes to do still and I just don't have the mental bandwidth to do it right now. I'm trying to do an OSS release of MIRA by the 31st and thats never going to happen if I'm trying to build a whole new protocol lol.